Fix auto-merge in downstream instances
We had a bug where the automerge functionality tried to use `pusher.py`
before `usrc.py` had run. This did no cause any issues in our testing,
but for jenkins instances that use their own git repo for configuration
which uses `usrc.py` to get code from this repo, this cause the
automerge functionality not to work at all.

This patch changes operation ordering so that `puser.py` is not
attempted to be used before `usrc.py` had run.

Change-Id: I287093c4915b5e18a23ea30bdeeb6b412697dc42
Signed-off-by: Barak Korren <bkorren@redhat.com>