Adapt for lowercase GitHub headers
A recent change in the generic webhook plugin made GitHub HTTP headers
be sent in lowercase, as opposed to mixed case which have been used so
far.

A fix to the `standard-webhook-trigger` job had already been submitted
and merged. This patch fixes the `standard-stage` job.

Also fixed a code issue that caused the `enqueue_change` function not to
be properly called occasionally.

Change-Id: I4ac001fce29ffa991ac5a06dbd41c6b1ad1a180b
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1923
Signed-off-by: Barak Korren <bkorren@redhat.com>