Workaround for a race in gerrit-trigger plugin
When reporting back to Gerrit, env vars are being evaluated from the shell of
the 1st job that was triggered. Since we use
GERRIT_TRIGGER_CI_VOTE_LABEL for custom messages, we hit this race
condition in some cases. This patch sets an empty string to
$GERRIT_TRIGGER_CI_VOTE_LABEL in case where standard-enqueue is the 1st job to
be triggered for post-merge events.

Change-Id: I3761bf9f911020eb88be38ea04ebed3e42294a73
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>