Simplifications and fixes to mock_runner.sh
The main body of work here is removing mock_runners support to loop
through several scripts and environments in a single run. Nobody ever
used this and it just complicated the code.

As a result of the work above, we now know that mock_runner.sh always
runs a single script in a single environment. This allowed for
simplification of the logs directory generated by mock_runner.sh and
making its structure completely predictable and static.

Several other small fixes are in place as well:
- One no longer needs to have the environment name immediately follow
  the '--shell' option
- Fixed an issue with the code trying and failing to handle logs in
  cases where no logs were generated.
- Made it unnecessary to specify the '-C' or '-mock-confs-dir' options

Change-Id: Ie6de374e58665a197a7d920f2af29cd716f7b8f8
Signed-off-by: Barak Korren <bkorren@redhat.com>