checkout_repo to tag every build
In order to create a pipeline based poll upstream sources job, we need
to tag every build so pusher.py will have a reference if it needs to
push the change or not (passed to 'push --unless-hash').

This patch adds the required functionality of tagging to checkout_repo:
1. Configure git user name and email.
2. Tag every build.

To make the git user name and email configurable, they are injected
via EnvInject plugin.

Change-Id: I4c67989b66cc34bc6e4fc42887c30e853206694c
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1955
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>