Introduce STDCI summary
The report is being rendered at the end of the build and provides a
simple UI for users to view the status of each STDCI thread that ran
during the job as well as it provides a convinient access to per thread
log and artifacts.

There are also lins to retrigger the build, easy access to junit
results/findbugs results (if there are any).

Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2050
Change-Id: I62f6c1675a52bbaa4519d33619c4f984c15c8800
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>