Fix mock_cleanup.sh trying to umount bad things
`mock_cleanup.sh` had some fail safe code in it to prevent it from
trying to umount things that cannot be un-mounted. Only that code had a
bug in it the caused it to do the exact opposite of what it was supposed
to do.

Change-Id: I0645fa1ca52054fd3effcccad177dbe096676acb
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1975