Update slave mirror snapshots
This is done to get an updated version of mock.

This was done using an automated tool so slight syntax changes were made
to all lines. I decided to keep thos changes so that we don't get the
same changes the next time we use the tool.

Change-Id: I50e234dc6fdd8a537526edf4c0cfeb04a25b5332
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2190