std stage: Write log to stdout instead to a file
At the point when pusher.py is invoked in std stage, we still don't have
the proper file structure to write the logs so it fails. With this
patch, pusher.py will write it's log to stdout so we'll be able to debug
it if needed.

Also set pusher.py to run in debug mode.

Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2236
Change-Id: I132083fdb23fd88a92c0134b1b27010fe1507636
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>