test_nested_config: removed unsupported test case
The dedup method assumes that the iterator of vectors it receives can't
be empty but still we used to test this case. With latest python3
update, the behavior of iterators has been modified and now they raise a
RuntimeError when calling next() on an empty iterator (used to raise
StopIteration).

For more infomration about the change:
https://www.python.org/dev/peps/pep-0479/

Change-Id: I2c89b59abf7cb6d25eedd5c011c3af8b2f9d7da4
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>