docker_cleanup: fix already removed container race
This patch fixes a possible race condition when removing/stopping
containers. Since we create a list of all containers and then we iterate
over this list to stop+remove them, it's possible that some containers
will be removed by the time we try to remove it.

Change-Id: Ifc950ce574fdc6d6f0e7e5fe3bd2d38390e9f566
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2314
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>