STDCI: Archive artifacts earlier
Instead of stashing STDCI artifacts when an STDCI threa is done and then
archiving all of them together, archive the thread's artifacts
immediately when it is done.

This will enable analysing thread issues earlier without having to wait
for all other threads to finish. As a benefit it also simplifies the
code and opens the door to removing the need for allocating an
additional node when the threads are done.

Change-Id: Iea9aa66867c3f11c36d8dfaf0ddc50a8414e1853
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2343