mock_runner: don`t store logs in source checkout
mock_runner.sh used to keep its logs in a temporary directory inside the
source checkout. This made it easy to move the logs to and from the mock
environment, but some projects turned out to have build/test scripts
that can damage the logs, therefore moved the temporary storage location
for the logs to /tmp.

To move the logs out of the mock environment, they are first moved out
from /tmp into a temporary location in the source checkout (this is done
after the build/test script had finished running so it cannot harm the
logs), and then moved from there to the temporary location used by
mock_runner outside of the mock environment.

Change-Id: I4a1eed70e8464a64cc07b256687bbd1385f7ca42
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2386