standard-enqueue: improve Gerrit message for V2
When standard-enqueue cannot figure out to which change queue should a
project's artifacts go, because there are no build-artifacts jobs for
that project, it sends a warning to Gerrit letting developers know their
patches are not being sent to the change queues and the nightly repos.

This warning is issued even if the project was switched to STDCI V2 and
has a proper release branches configuration to send patches to the
change queues.

This patch modifies the warning text to explain that it is irrelevant
fro projects that use STDCI V2.

Change-Id: Ie4e2f118363a4a4e3c253fb9c6293ea8983e7fdc
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-ticket: