stdci_runner: don't fail when project setup fails
We're having sporadic failures during docker_cleanup in some of the
jobs where python requests reaching it's timeout when trying to delete
container images.

Instead of failing, the build will send an email to infra@ovirt.org to
notify the team. The infra owner should check that we're not running out
of disk space due to a large amount of container images left.

This is a temporary workaround to debug and follow up with this issue,
and by any means is not a permanent solution.

Change-Id: Ia75280efaf742a6e3f69d033e6520545a03d734a
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2504
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>