dropping infra@ovirt.org from specific ost suites
infra team isn't monitoring suites that are not part
of the CQ, and are maintained by the relevant maintainers.
Having these suites email infra@ovirt.org creates a very
large noise, especially if they are not handled by maintainers
promptly.

If a suite fails, the maintainer should reach out to infra
if help is needed or an infra issue is detected.
If a maintainer believes additional people/lists should be
added, please comment on this patch

Change-Id: I96ed4064904d45a6d5ef6d9a03705497eb4c014f
Signed-off-by: Eyal Edri <eedri@redhat.com>