Make usrc.py able to get files to remote branch
upstream_sources.yaml could be extended with dest_format parameter
which accept list of values. Currently recognized values are
'files' and 'branch'. 'files' is the default option (if dest_format
is not specified) which gets the upstream commit as files. If
'branch' exist as value the upstream commit is pushed to downstream
remote branch. The name of the branch is '_upstream_' prefix, the
name of the branch from upstream_sources and the short commit sha,
e.g _upstream_master_731a26c.

Change-Id: I4ac916964e30674da347aa0b638e04e94c24fcc2
Signed-off-by: Emil Natan <ena@redhat.com>