slaves_deploy: Added auto-config of kube-plugin
Added code to the slaves_deploy playbook to create ConfigMap objects
that get automatically imported into Jenkins by the OpenShift-sync
plugin as slave pod specifications for the kubernetes-plugin.

Each kind of slave needs to be configured as a host in the Ansible
inventory file and include a `podspec` variable that contains the YAML
specification for the slave pod.

Jenkins configuration details such as the Jenkins labels for the slaves
are configured by setting specific labels in the pod specification.

Change-Id: I28fd765df0011f9f9287b1d8d37528755b5bc7c4
Signed-off-by: Barak Korren <bkorren@redhat.com>