mirror_client: fix fc29 mock to use mirrors.
In fc29 we need to change the 'no proxy' value we inject to the yum/dnf
configuration to 'None' instead of _none_.

We made the mirror injection code use the repo name suffix to decide
what value to use. We also made it possible to explicitly specify the
value and made the specification explicit in the mock configuration
files.

Change-Id: I8abbcb2c4d25c097bfdad7834928a45f405e0d39
Signed-off-by: Ehud Yonasi <eyonasi@redhat.com>
Signed-off-by: Barak Korren <bkorren@redhat.com>