cleanup_slave: Remove old engine data
We seem to have `/var/lib/ovirt-engine` directory on the slaves that
takes quite a bit of space. This most probably dates back to the times
when we ran upgrade jobs outside of mock.

This patch adds functionality to `cleanup_slave.sh` to remove old ovirt
directories from the slaves.

Change-Id: Ie9ef8850d530f86ea2ca8b0946408457505d3f3c
Signed-off-by: Barak Korren <bkorren@redhat.com>