Containerized loader node
Instead of using a VMs as general purpose node (loading code, rendering
build summary and etc), we will use a loader-container which is a simple
non-privileged container that runs in an OpenShift pod.

Since this container is pre build for this mission and it is
ephemeral, we no longer need to run global_setup and slave_cleanup.

Change-Id: I6b8f11463a95a611da21d79afb625aceb0901b12
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>