mock_runner: handle server errors in try_proxy
try_proxy() is used to detect whether we can use proxy. If we can't we
automatically fallback to the default non-proxied config. Currently, we
only handle timeouts but not other http errors such as permission
denied.

It's needed now because we're adding external hosts that can't and
shouldn't access our proxy in phx.

Change-Id: I34d21ae57728b30009d44c272e06e70c6f8396fc
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>