master_deploy: Disconnect build from deployment
The master_deploy playbook used to look at updates for the master
container base image in DockerHub and the master container source code
and built it if any changes were found.

If a new image was build, the playbook would try to deploy it
immediately there by causing running master to get restarted.

This patch disconnects the link between building and deploying so that
the image that is deployed is hard-coded and does not change until it is
updated explicitly in the playbook source code.

For the time being the build process is still embedded into the overall
master_deploy playbook, but might be moved away going foreward.

Change-Id: Ia35981f3fbe3541f7146e6eac4747e2c842041be
Signed-off-by: Barak Korren <bkorren@redhat.com>