mock_runner: Fix stuck `tee` processes
Even though we have a timeout defined in `mock_runner`, we occasionally
see `mock_runner` processes running indefinitely with a `tee` process
running inside the mock environment and keeping it alive.

Te reason for this is that our code is built in a way where is waits
indefinitely for the `tee` process to exit. In the situation we`re
seeing, the user script starts up a background process that ignores the
`kill` signal from out `timeout` command while still holding a file
descriptor attached to the `tee` command thereby preventing it from
exiting.

This patch changes the way we start up `tee` so that it runs in the
background and our code no longer relies on it existing.

We've already verified that if the `tee` command sticks around `mock`
can successfully kill it as we exit the mock environment.

Change-Id: I33b4c61867dab4f0030afcdf66d4a1533b01af2e
Signed-off-by: Barak Korren <bkorren@redhat.com>