mock_runner: Seamless SSH from mock_runner
Openning SSH connections out of `mock_runner` takes some work. While
getting access to authentication keys had been sorted out by earlier
patches. One still had to take care of other aspects of establishing an
SSH connection including:

1. Taking care of or bypassing the `known_hosts` file
2. Figuring out the username to use for the connection

The `known_hosts` file is handled by keeping a copy of the file in the
`jenkins` Git repo and then placing it in the right place withing the
`mock_runner` environment.

The username is handled by generating an `~/.ssh/config` file in the
`mock_runner` environment that sets the default SSH username to either
be the same name as the user that invoked `mock_runner` or the name
defined in `SSH_AUTH_USER` if it is defined in the environment
`mock_runner` is launched from.

To implement the above, the mechanism that handled the `ci_toolbox`
directory had been generalized to allow bind-mounting copies of
arbitrary directories from the `jenkins` source repo. This mechanism was
then used to mount the following files/directories:

1. The `data/ssh_files/` directory was bind mounted to make the
   `known_hosts` file available.
2. The `scripts/mock_runner_profile.sh` file was bind mounted to
   `/etc/profile.d` to provide an easily customizable startup script for
   the `mock_runner` environment.

The `mock_runner_profile.sh` script takes care of creating or copying
all the SSH configuration files and setting the right permissions to them.

To make the `mock_runner_profile.sh` script run when the environment
starts up, `mock_runner` had been made to pass the `-l` parameter to the
`bash` shell that is invoked inside `mock`. This makes it run the
various scripts in `/etc/profile.d/'.

Change-Id: I1ffb55002a08bfb6d640972c8789fd17d98c8960
Signed-off-by: Barak Korren <bkorren@redhat.com>