build_images: Added playbook
Added a playbook for building container images in OpenShift.

The playbook is meant to be run after the `import_base_images` playbook
had ran and imported base images as needed.

The build images are uploaded into `quay.io` (by default) and tagged
with tags that include the image source and base image checksums. An
additional tag can optionally be added to mark the latest build of an image.

Currently only the s2i build strategy is supported, support for other
strategies is planned to be added in the future.

Change-Id: I2835fae53223445b0350691370145584ffd62436
Signed-off-by: Barak Korren <bkorren@redhat.com>