update_images: Add playbook for image updating
Added the `update_images` playbook that looks for updated container
images in quey.io, and writes their tags to local files that would be
used by other playbooks.

Also including the results of running this playbook in this commit as
well as the files need to invoke it from the STDCI
`poll-upstream-sources` stage.

Change-Id: I62ce3b1253871596fc91c2e729e05d96238856f5
Signed-off-by: Barak Korren <bkorren@redhat.com>