mock_runner: Fix fcraw environment
Fix issue with `fcraw` environment where it was failing to install
`docker` because it is no longer shipped in Fedora.

This patch installs Docker Community Edition in the mock environment.

This patch is likely to still fail CI because of an issue in mock itself
where it fails storing a log file after creating the mock cache. The
following PR would fix that issue:

https://github.com/rpm-software-management/mock/pull/252

We must wait until that PR is merged and a fix version of mock is
released.

Change-Id: I8e0064de7dedc6ad7a1c0d57812a760ed1051af1
Signed-off-by: Barak Korren <bkorren@redhat.com>