mock_runner: Don't define env vars during init
The commands that run during `mock init` have some rather aggressive
logging that also includes all the environment variables that are
defined when they run.

This has caused secret values to be dumped to the root.log file that is
generated during the 'init' phase.

This patch changes `mock_runner.sh` so that it defined the environment
variables only after init is already done.

Change-Id: Ie513ef392f885a8bc61ee1689d053ba622a7cad3
Signed-off-by: Barak Korren <bkorren@redhat.com>