usrc.py: add support for dest formatters params
Refactored the dest formatters handlers and the get
method of GitUpstreamSource so that dest formatters
will be able to get params from the configuration file.

Each dest fromatter needs to fulfill an API and accept
**kwargs in it's params. It is then free to decide if
it wants to use the params or not. User can specify in
the configuration file the values for the arguments of
specific formatter, and those args will be passed to it.

Change-Id: I631466e0e188177db7a14acb9d4560daacf2b641
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>