usrc: Use `anonymous_clone_url` from push map
The `source_repos` formatter for `usrc` used the `push_url` from the
push map file as the URL specified in the `source-repos` file it
generated. This URL turned out to not comply with the requirements of
various Koji instances. We've now added an optional
`anonymous_clone_url` field to the push map file to be used for this
purpose instead.

The value of the `anonymous_clone_url` defaults to the value of
`push_url` if it is unspecified in the push map file, thereby
maintaining compatibility with current functionality.

Change-Id: I7b354d2af4b529819dc1ef8a08656794f06117b4
Signed-off-by: Barak Korren <bkorren@redhat.com>