usrc: allow choosing push url from push map
In order to push different sources to different remotes,
we now allow choosing the remote url from the push map.
Also, we can propogate the call from branch format handler
to source repos format handler. This functionality isneeded
in order to let users to choose the remote url once instead
of having to specify it twice.

Change-Id: I5f7b976a6e57c44b59af36faa15d1f9920f45bc6
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>