gate: Filter releases by change queue prefix
Add a JJB parameter to the gate job template where the system name for a
set of queues should be given (A name such as 'ovirt' or 'rhv'). The
gate job will, when extracting release queues from the
`release-branches` section of `stdci.yaml` for patches, filter out
queues that do not belong to that system, and will then strip out the
system name from the queue name, so that the list becomes a list of
system versions.

For example, if a project targets the 'ovirt-master' and 'ovirt-4.3'
queues in its `release-branches` definition, and the `st-queue-prefix`
parameter is set to `ovirt`, the system will know it needs to run suits
for the `master and `4.3` versions of oVirt.

If a patch for the same project above is sent into a gate job where the
`st-queue-prefix` is set to `rhv`, the list of release returned for it
would be empty, and the patch would effectively not be tested with this
gate.

Change-Id: I124b5823ea8a1ceabda5b78fb86ba99f6eaeb74e
Signed-off-by: Barak Korren <bkorren@redhat.com>