usrc: Stop processing tags if the tag list empty
After we filter the tags, we try to find the newest one using the max
function. If the list of tags is empty, max will raise an exception.
If this is the case, there is no new tag that we can consume, so return
the commit that was already in the upstream-sources file.

Change-Id: I47760dc4cd66802ba1a2f63bfcc1b52a578f4aa7
Signed-off-by: gbenhaim <galbh2@gmail.com>