usrc: Adjust _rev_parse to return an ascii string
In usrc, we assume that strings that represent a commit are of type str.
In Python 2, the return type of the stdout of subprocess can be
'unicode', if this is the case, we need to encode it to ascii, so it
will be of type str. Since the commit contains only HEX chars, it is
safe to encode it to ascii.

Change-Id: Ic739c150f791c0431e3f5dcba968b6caaf4960e3
Signed-off-by: gbenhaim <galbh2@gmail.com>