gate: Enable gating of ST project
Modify the gate job so that if it encounters a patch for the project
that it gets the test suits from, it will clone that patch instead of
cloning the master branch.

This will effectively allow gating of projects like OST by sending their
patches through the same Zuul pipeline and change queue as the patches
of the projects they test.

Change-Id: I064bc3891458bae516b07eec95c01acd871d98b5
Signed-off-by: Barak Korren <bkorren@redhat.com>