standard-deploy: added OpenShift-based deploy job
Added a job for deploying packages to containers in OpenShift so that
OpenShift end up resembling a yum repository server.

The new job takes over the name of the old experimental deployment jobs
to avoid having to add a new regex to `pipeline-loader.groovy`.

The new job leverages our existing ability to launch playbooks via
STDCI, therefore several `automation/*` files are added to allow this.

Change-Id: Ia8bc39be61dcff7e3c5bb058eb6917e99edd4ad6
Signed-off-by: Barak Korren <bkorren@redhat.com>