gate: Wait for post-merge jobs
Make the gate pipelines wait for builds of post-merge jobs that started
before the gate build started and are going to deploy artifacts into a
gated repo that would be used by that gate.

This would prevent a race condition where the gate can run tests that
will not be able to access a build that was made from a patch merged in
a previous run of the gate.

Change-Id: I700f3e28cc6b6a5586a9fca9007ab064da05c084
Signed-off-by: Barak Korren <bkorren@redhat.com>