mirror_client: Support repo insertation
Extend the mirrors mapping syntax to support having extra repos be
added into the resulting yum configuration before certain repos if those
repos appear in the configuration.

For example, we can specify that the `gated` repo should be added before
the `tested` repo in any file that that includes the `tested` repo.

Change-Id: Iddaff9a4e2fc6927d9426b22bc0e8a425862da2f
Signed-off-by: Barak Korren <bkorren@redhat.com>