mirror_client: Add mirrors_to_inserts function
Added a function that allows converting a mirrors data map into a
different map where all mirrors in the map are set to be inserted before
other given repos.

This is meant to be used to convert the mirrors data downloaded from the
'gated' repos into insertions before the 'tested' repos.

Change-Id: I33ed6b8414e63fc3448a5aee68bd54a9cca1ec66
Signed-off-by: Barak Korren <bkorren@redhat.com>