mirror_client: Support mirror includes
Add syntax to mirrors data files so they can include other mirrors data
files.

This is done so we can create mirror files that join multiple repo
sources together, an example use case for that is injecting both the
'regular' mirrors and the gated repo at the same time in the gate job.

Change-Id: Icbb256b7384669866a05804ba0bff3dc14e880e3
Signed-off-by: Barak Korren <bkorren@redhat.com>