gate: Insert gated repo via mirror injection
Have the gated job setup a mirror injection file in such a way that
the injection process would insert the 'gated' repos into yum config
files that include the 'tested' repos as well as inject the usual set
of mirrors.

This patch relies on the inclusion and insertion capabilities that were
recently added to `mirror_client.py'.

Change-Id: I682fcfc032ad8cb8c8d58d358019070b825351b1
Signed-off-by: Barak Korren <bkorren@redhat.com>