dummy_suite_master: Check for gated repo
Add a check to the `dummy_suite_master` gating test suit for the
insertion of the `gated` repo into `*.yumrepos` files.

This patch requires the following patch to be merged or applied with JJB,
and for the resulting job to be invoked for the relevant repo to exist
in the 1st place:

    https://gerrit.ovirt.org/102553

This patch also requires that the patch its based on will be merged for
the gate tests to pass, because merging will cause a package to land in
the gated repo ensuring that it is not empty.

Change-Id: I686159c94e83311b2aaaf26ed2286c28ac0c672f
Signed-off-by: Barak Korren <bkorren@redhat.com>