update default git clone timeout for pipeline jobs
Due to slow network from Gerrit to PHX, we're getting
errors to clone git repos and the default 10 min timeout
isn't enough.

We've updated the timeout for V1 jobs, via yaml templates
but for pipeline jobs using v2 it needs an update to the
pipeline groovy code.

This patch updates the plugin default 10 min to 20 min
in the checkout function which should affect all the pipeline
jobs in CI

Change-Id: If6ffbe95e19ad55f6b3bca5b07153b447c2fd60a
Signed-off-by: Eyal Edri <eedri@redhat.com>