global-setup: Fix Python package names on Fedora
Fedora had moved to unified Python package naming where all package
names include the Python version they are meant for. In FC30 some
packages have also started removing the `provides` mentions that
included the older names, so using the older names no longer works.

This patch ensures we use the newer package names in Fedora.

This patch also removes the `remove_packages` function, it was used to
fix a one-time issue in EL7, and is now significantly slowing down setup
times on Fedora.

In addition to the above:

- Since FC30 no longer includes the Docker API for Python2, we move to
  run `docker_cleanup.py` on Python3 across the board, and use the
  appropriate API.
- We include a workaround for the following FC bug:
  https://bugzilla.redhat.com/show_bug.cgi?id=1724305

Change-Id: I1e41106dd6720ff01018be7540f12a733458678d
Signed-off-by: Barak Korren <bkorren@redhat.com>