pipeline-loader: Refactor library loading
Re-factor library loading in pipeline-loader so it works more like
importing in Python and does not try to load the same module more then
once.

We also remove the ability for libraries to inherit functions from the
files that load them. Inheiting from pipeline-loader itself is still
allowed though.

Doing this allows removal of most of the loading builerplate code we had
to include in all our libraries, and it makes sharing arbitrary
functions via libraries a lot easier.

This patch also modifies all the existing pipelines to use the new
loading code.

Change-Id: I92fb7e71c2fd92e9372f10683a481552b71b115a
Signed-off-by: Barak Korren <bkorren@redhat.com>