pipeline-loader: Enable groovy code self-test
Our approach for testing STDCI code itself so far was to have the CI
jobs for the STDCI repo (A.K.A `jenkins`) use the latest code with
current modifications included so that if the code is broken the jobs
would not pass.

Because of the way is was being loaded, our pipeline Groovy code was not
tested in this manner. This has led to us having to perform extensive
tests to Groovy code manually, or otherwise find issues that can affect
the whole system only after the code had been merged.

Recent failure events have convinced use that an automated approach for
this is critical.

This patch makes the pipeline-loader detect if the job is in fact
testing a patch to the STDCI repo. If that is the case, the patch code
would be loaded as opposed to the code from the master branch or the
refspec configured by JJB.

With this change, the only piece of code that is not being loaded from
the patch when the job runs is the loader code itself.

Change-Id: Iae6e955b11ce8a5477a1077039212a92d2d4d6fb
Signed-off-by: Barak Korren <bkorren@redhat.com>