gate: pass change attributes to Jenkins job
Make the gate Zuul job pass values to the Jenkins job parameters as it
triggers it

This will make it easier to know which change every build of the Jenkins
job is related to which change in Gerrit.

As a side-effect of this, if the gate job is testing a change to the
`jenkins` repo itself, the changes we already introduced to the
`pipeline-loader` script would make it load the updated version of the
gate pipeline from the tested change. This means that the gate job can
be used to effectively test its own code without having to manually
apply the new code via JJB/Jenkins GUI.

To be tested and merged successfully, this change requires that that the
change that precedes it (For adding parameters to the gate Jenkins job)
would be successfully merged and applied via JJB. Otherwise the Zuul job
would fail to trigger the Jenkins job.

Change-Id: Ia3eac404014d85c9e90e1e51deea48aa483cbc42
Signed-off-by: Barak Korren <bkorren@redhat.com>