gate: Get change attributes as Jenkins job params
Make the gate Jenkins job accept parameters that contain details about
the change being checked by it.

If a future patch we will make the gate Zuul job fill-in these
parameters, it will make it easier to know which change every build of
the gate Jenkins job is related to which change in the SCM system.

Change-Id: I90f6e73d327465d8c5e093d29d54dd2ba45b149c
Signed-off-by: Barak Korren <bkorren@redhat.com>