JJB: Save STDCI Git SHA statically in jobs
All of our jobs clone the `jenkins` repo to obtain the CI infrastructure
code to use. So far, the jobs had been cloning the repo's `master`
branch. That could have cause issues in cases where the job
configuration needed to be updated with JJB before it could use the new
code.

The patch modifies the JJB update process so that it stores the Git SHA
of that code that was used to configure it in the job configuration so
that it does not try to use newer code unless it is updated with JJB.

Since we now have to update the jobs every time our code changes,
`stdci.yaml` was changed to do so.

Change-Id: Ia968279d6f66599c20db48c27795d39e10ae5c3e
Signed-off-by: Barak Korren <bkorren@redhat.com>