Adda an STDCI DSL option for using containers
Add the `containers` option to STDCI DSL, this option would eventually
allow to replace `mock` with containers for managing build dependencies.

In an attempt to make the DSL options module more modular, the option is
implemented as a class, some refactoring is done to enable that.
Additionally a normalization mini-DSL was implemented to allow a more
declarative implementation of option classes.

Change-Id: I662759af0cc54baa88644e409a86e636bcdde914
Signed-off-by: Barak Korren <bkorren@redhat.com>