STDCI PODS: Unique UID for each job build's POD
Configure the STDCI PODs so that all the pods that belong to the same
job and build run with the same UID which is also different from the
UIDs of PODs for all other jobs and builds.

The PODs UID is derived from the job name and build number with some
random data added. Since job names in STDCI include project ans stage
names, this would guarantee that PODs of different projects and
different stages will get different UIDs

Since the PODs now self-select their UIDs, we add the serviceAccount
setting to then so they have enough permissions to do so.

Also removed some code that tried to implicitly set PODs to run in
privileged mode. It was setting this in the wrong place and should be
done explicitly anyway.

Change-Id: I8a5777248b182dde3d7833966ea7296526eece1f
Signed-off-by: Barak Korren <bkorren@redhat.com>