pipeline-loader: Improve change detection
Improve change detection for the pipeline-loader script - now detect
when change happened in a previous unmerged commit, and not just the
latest one.

This helps with having CI work properly when having chains of patches
where later patches require changes to `pipeline-loader.groovy` but do
not include such changes themselves.

Change-Id: I8e58e069a9d482c58830687b2a8f102328383d22
Signed-off-by: Barak Korren <bkorren@redhat.com>