Add dockerfile_utils.py
The dockerfile_utils.py script will include utilities for managing
Dockerfiles.

The first utility that was added is 'parent_image_update' that can be used
to automatically update the image reference in a from command, base on a
floating tag.

More information about the usage can be found in the docstrings and the
CLI help.

Change-Id: Ib7930acf9993607035e7f3a00c1427c184133365
Signed-off-by: gbenhaim <galbh2@gmail.com>