usrc.py: allow manual updates of upstream sources
Some of our users have their own mechanism of identifying the
value of the commit of their upstream sources and can't use
our built in update policies.

This change allows such users to explicitly specify an upstream
source entry to set in the config using the modify-entries command.
The command accepts an IFS separated upstream source entries in a
form of YAML string.

Upstream source entries that already existed in the config will be
replaced at their original location. New upstream sources will be
appended to the end of the config.

Change-Id: I8926a83f90ffc0a12ad89eaf3667accb50e7bc04
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>