pusher.py: Avoid re-pushing closed changes
`pusher.py` has a built-in check to find out if a patch it is going to
push had already been pushed before, and avoid pushing it if that is the
case.

That check was so far only looking at the open patches in Gerrit. This
made it return a false negative if the patch was already in Gerrit but
was abandoned or closed in some other way. This made `pusher.py` try to
re-send it and subsequently fail when Gerrit rejected the new patchset
on grounds of the patch being closed.

This patch makes `pusher.py` look for its already pushed patches among
the closed patches as well.

Change-Id: I162a610d174ef0d736d291b7cda0b6f3e33d0d65
Signed-off-by: Barak Korren <bkorren@redhat.com>