pipeline-loader: refactor: Use podspec struct def
Switch from using the YAML-string-based PodSpec definition style for the
K8s plugin to the object structure style, as the latter allows to use
some more plugin features such as PodSpec inheritance that we're going
to need soon.

Note that the some features (Notably `runAsUser` and `runAsGroup`)
require kubernetes-plugin v1.21.3 or newer. With older versions this
patch may slightly reduce the system security.

Change-Id: I2bf8d63594fd2d1a58452f3011db7ae455a6381c
Signed-off-by: Barak Korren <bkorren@redhat.com>