STDCI PODs: Add artifact collection logic
Let the CI system perform artifact collection for PODs of threads that
have the `decorate` option.

This is done by having the loader node run an NFS server that exports a
shared space that is then mounted on the PODs in a way that allows then
to put their artifacts in it.

Once the PODs are done running the CI system uses the agent running on
the loader node to collect the artifacts from the shared space.

Change-Id: I6c8b98d64bb8819f3f5b9cb83bb8159af0cbe555
Signed-off-by: Barak Korren <bkorren@redhat.com>