gate: Support getting suits from STDCI DSL
So far the gate job generally got suits to run by scanning the test suit
repo (E.g. OST) for suit scripts. It then hardwired all the other
runtime parameters around those scripts to convert them to CI threads.

With this patch, the gate job now also looks at the substages of the
`gate` state in the STDCI DSL YAML file. If substage names match names
of suits that would be run by the gate, the job will no run those
substages instead of running the suit script directly.

This allows this job to take advantage of new features in the STDCI DSL
such as the container support.

Change-Id: Ic4fd7006def9fdc7d5168a81d8ed989a94b538fc
Signed-off-by: Barak Korren <bkorren@redhat.com>