pusher,usrc: Fixed argparse issue on Py3
It seems there is an issue with argparse in Python 3.6 where an error is
not raised if the name of a subcommand is not specified by the end user.
This leads to the script crashing elsewhere and needed arguments are
missing.

The following Python bug provides more info on this issue:

    https://bugs.python.org/issue9253

This patch includes a work around mentioned in the bug comments.

Change-Id: I3c6954493aaa665662f99a80e7de00205a91836f
Signed-off-by: Barak Korren <bkorren@redhat.com>