Adding Pipenv configuration & setup.py
[Pipenv][1] is a modern Python dependency management tool, that makes it
easy to manage code dependencies in a reproducible way.

`setup.py` is the main configuration file for the Python setuptools
package.

Together these tools allow us to:
1. Setup a dev/test environment with a sinlgle `pipenv sync` command
2. Install our scripts (usrc, pusher, etc.) as executables in $PATH
3. Track dependencies in a reproducible way

The added configuration runs some of our tools (usrc, pusher, mirrors)
and with Python3 in a pure virtual environment without access to the
system Python packages. This makes it much easier to run the tools in a
reproducible way across distribution versions.

This patch moves the python3-only tests to use `pipenv` for dependency
management instead of the older stdci-venv script. It also adds a
`*.mounts` file and an ~*.environment.yaml` file to place caches locally
on the slave.

We may desire to move everything to work with Pipenv like this, but in
some cases we have code that runs inside other systems (E.g. mock) where
we cannot have full control of the Python environment, and therefore we
must keep testing some parts of our code against system packages.

There are Python-related things in this repo that are not yet supported
by the Pipenv environment - those include `jjb` and the playbooks, Right
now the focus was on running the kind of tools we want to have running
inside a container in the near future.

[1]: https://pipenv.kennethreitz.org/en/latest/

Change-Id: Ieefa6a08929440703dd58af258f790d1fea69988
Signed-off-by: Barak Korren <bkorren@redhat.com>