Added a container with STDCI tools
Added files for building a container image that contains several useful
STDCI tools, including `usrc.py`, `pusher.py` and others.

The container is based on the RHEL8 UBI Python 3.6 image and contains a
preinstalled Pipenv environment that includes the requirements for
running the various tools.

The container is meant to be used as an initContainer for providing
STDCI capabilities for PODs running build and test workloads, as well as
provide a convenient way to gain access to STDCI tools.

Change-Id: Icd2ed847b7aabcbf57bfbecaaccd39176a46bf12
Signed-off-by: Barak Korren <bkorren@redhat.com>